requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
openai==1.75.0
plantuml==0.3.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-xz==0.5.0
rdflib==7.1.4
requests==2.32.3
shexer==2.6.0
six==1.17.0
sniffio==1.3.1
SPARQLWrapper==2.0.0
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
wlighter==1.0.1
